+++ WDG_ex.c
@@ -46,6 +46,10 @@
 #endif
 
 
+static const uint32_t refCLK = LSI_VALUE / 1000;	// Watchdog clock frequency in KHz (for 32b computations)
+static const uint32_t prescaler_mult = 4;			// STM32 IWDG STM32 architecture prescaler multiplier
+
+
 #if defined(HAL_IWDG_MODULE_ENABLED)
 static bool iwdg_en = false;
 static IWDG_InitTypeDef IWDG_cfg = { 0 };
@@ -227,14 +231,12 @@ HAL_StatusTypeDef WDG_restore_cfg(void)
 #if defined(HAL_IWDG_MODULE_ENABLED)
 HAL_StatusTypeDef NONNULL__ set_IWDG_Period_us(IWDG_HandleTypeDef * const pIwdg, const uint32_t per)
 {
-	const uint32_t refCLK = LSI_VALUE / 1000;	// Clock frequency in KHz (for 32b computations)
-	const uint32_t min_prescaler = 4;			// STM32 IWDG STM32 architecture minimum prescaler
 	const uint32_t max_prescaler = 256;			// STM32 IWDG STM32 architecture maximum prescaler
 	const uint32_t max_reload = 4095;			// STM32 IWDG STM32 architecture maximum reload value
 	uint32_t reload;
 	uint32_t prescaler;
 
-	for (prescaler = min_prescaler ; prescaler <= max_prescaler ; prescaler <<= 1)
+	for (prescaler = prescaler_mult ; prescaler <= max_prescaler ; prescaler <<= 1)
 	{
 		reload = (refCLK * per) / (1000 * prescaler);		// Computation using KHz instead of Hz to stay in 32b range
 		if (reload <= max_reload)	{ break; }				// If in range
@@ -244,9 +246,17 @@ HAL_StatusTypeDef NONNULL__ set_IWDG_Period_us(IWDG_HandleTypeDef * const pIwdg,
 	if (prescaler > max_prescaler)	{ return HAL_ERROR; }	// If nothing has been found (after last iteration)
 
 	pIwdg->Init.Reload = reload;
-	pIwdg->Init.Prescaler = (prescaler / min_prescaler) - 1;	// Convert to register value (0-7)
+	pIwdg->Init.Prescaler = (prescaler / prescaler_mult) - 1;	// Convert to register value (0-7)
 
 	return HAL_IWDG_Init(pIwdg);
+}
+
+uint32_t NONNULL__ get_IWDG_Period_us(const IWDG_HandleTypeDef * const pIwdg)
+{
+	const uint32_t reload = pIwdg->Init.Reload;
+	const uint32_t prescaler = (pIwdg->Init.Prescaler + 1) * prescaler_mult;
+
+	return ((reload * prescaler * 1000) / refCLK);
 }
 #endif